* Alterado o objeto de D/A para permitir fornecer o valor direto, em milivolts ou porcentagem. corrigido #817 + Adicionado campo de ID do ModBUS da POP-7. corrigido #760 * Alteradas cores de fundo e do texto do listview da janela de Debug. corrigido #818 * Atualizado o help para o formato CHM (Formato em HTML da Microsoft), restando adicionar conteudo mais relevante. ticket #39 + Adicionados diversos novos icones que representam os comandos no Ribbon. ticket #14 * Corrigido objeto SetDA que declarava um valor numerico na lista de I/Os. corrigido #822 * Corrigido problema onde o ladder ficava inutilizavel quando ocorria erro ao abrir arquivo. Agora, se ocorrer um erro, cria um novo diagrama. corrigido #823 * Atualizada logica de navegacao pelo teclado. Agora procura por objetos mais proximos ao inves de ir em linha reta na direcao da busca. corrigido #18 * Corrigido problema de navegacao no ladder pelo teclado quando estava em um comentario. corrigido #26

+++ multisetdadialog.cpp
@@ -220,7 +220,7 @@ void CalcGainUp(int calc_time, int calc_initval, int calc_gainx, int calc_gainy 
 			VarDA = calc_initval - (DeltaYGanho * (calc_time - tm)) / DeltaXGanho;
 		}
 
-		gains[i] = VarDA+2048;
+		gains[i] = (float)(VarDA+2048);
 	}
 }
